src/todo/backend/settings.py
from decouple import config

SECRET_KEY: str = config("SECRET_KEY", default="changeme")
ALGORITHM: str = config("ALGORITHM", default="HS256")
ACCESS_TOKEN_EXPIRE_MINUTES: int = config("ACCESS_TOKEN_EXPIRE_MINUTES", default=30, cast=int)
