pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafel-Melo",email = "rafael.vilas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)"
]

[tool.poetry]
packages = [{include = "todo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
